requirements.txt

components>=0.0.1a0
pandas>=2.2.3
plotly>=6.0.0
requests>=2.32.3
streamlit-pdf-viewer>=0.0.21
streamlit>=1.43.0
trafilatura>=2.0.0
twilio>=9.4.6
pypdf2>=3.0.1
pdf2image>=1.17.0
streamlit-pdf-viewer
trafilatura
openai
trafilatura
openai
twilio
fantraxapi
numpy
plotly
pandas
streamlit
python-dotenv
requests
components
pandas
pdf2image
plotly
pypdf2
python-dotenv
requests
streamlit-pdf-viewer
trafilatura
twilio
